requirements.txt
# requirements.txt
flask==2.0.1
python-dotenv==0.19.0
ipaddress==1.0.23
